Day 5- Loop your world
Loop your world 
For this challenge, you need to take an integer value as input from stdin, calculate its factorial and print the result to the stdout. 1

Input Format
A single integer value to be taken as input from stdin and stored it in a variable of your choice. 

Constraints
1 < = n < = 15

Output Format
Print the value which you will get after calculating the factorial of the input. 

Sample TestCase 1
Input
5
Output
120