How Much Big Is Your Number
How Much Big Is Your Number 
For this challenge, you will take an integer input from stdin, store it in a variable and  calculate the number of digits in the number using division operator.

Input Format
A single integer value to be taken as input from stdin and stored it in a variable of your choice. 

Constraints
1 < = n < = 18

Output Format
Print the value which you will get after calculating the number of digits. 

Sample TestCase 1
Input
34567
Output
5