GCD of two numbers
GCD of two numbers 
For this challenge, you need to take input of two numbers , calculate their greatest common divisor (GCD) and print it to the stdout.

Input Format
You need to take two integers as an input which are separated by a single space. 

Constraints
1 < (a,b) < 10^5

Output Format
Print the single integer result to the stdout. 

Sample TestCase 1
Input
81 81
Output
81